Address review feedback: switch to Chat Completions, simplify sample
Per @eavanvalkenburg:

- Switch from OpenAIChatClient (Responses API, OpenAI-only) to
  OpenAIChatCompletionClient (Chat Completions). The sample now uses
  MLflow's unified /gateway/mlflow/v1/ endpoint, which works with any
  provider backend behind the gateway (OpenAI, Anthropic, Gemini,
  Mistral, Bedrock, Ollama, etc.) — not just OpenAI.
- Remove _require_env() helper; use os.environ[...] directly so it
  raises KeyError if missing. The docstring already lists required vars.
- Inline the streaming flow into main(); drop the non-streaming
  function so the sample is shorter and focuses on client setup.
- Rename file to mlflow_gateway_with_openai_chat_completion_client.py
  to reflect the client class.

Verified end-to-end against a running MLflow gateway: client connects,
auth passes, request is routed to the provider.
